backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
pyyaml>=6.0,<7.0
